src: reorganize c++ code structure
