+++ CMakeLists.txt
@@ -28,8 +28,5 @@ set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -Wall -Wextra -Wshadow")
 set(CMAKE_CXX_FLAGS_RELEASE "-O2")
 set(CMAKE_C_FLAGS_RELEASE "-O2")
 
-find_package(PythonLibs REQUIRED)
-pybind11_add_module(complatecpp MODULE src/complatecpp/module.cpp)
-target_link_libraries(complatecpp PRIVATE complate::quickjs)
-
-install(TARGETS complatecpp DESTINATION .)
+add_subdirectory(src/core)
+add_subdirectory(src/quickjs)
