+++ src/core/CMakeLists.txt
@@ -11,5 +11,16 @@
 #  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
-from .core import Value, Function, Stream, StringStream, Renderer
-from .quickjs import QuickJsRenderer, QuickJsRendererBuilder
+find_package(PythonLibs REQUIRED)
+
+pybind11_add_module(
+        core MODULE
+        core.cpp
+        ${PROJECT_SOURCE_DIR}/src/utils/mapper.cpp
+)
+target_include_directories(
+        core PUBLIC
+        $<BUILD_INTERFACE:${PROJECT_SOURCE_DIR}/src/utils>
+)
+target_link_libraries(core PRIVATE complate::core)
+install(TARGETS core DESTINATION .)
