+++ src/utils/mapper.h
@@ -1,0 +1,28 @@
+/**
+* Copyright 2021 Torsten Mehnert
+*
+*  Licensed under the Apache License, Version 2.0 (the "License");
+*  you may not use this file except in compliance with the License.
+*  You may obtain a copy of the License at
+*
+*      http://www.apache.org/licenses/LICENSE-2.0
+*
+*  Unless required by applicable law or agreed to in writing, software
+*  distributed under the License is distributed on an "AS IS" BASIS,
+*  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+*  See the License for the specific language governing permissions and
+*  limitations under the License.
+*/
+#pragma once
+
+#include <pybind11/pybind11.h>
+#include <complate/core/value.h>
+
+class Mapper {
+public:
+  Mapper() = delete;
+
+  static pybind11::object value_to_python(const complate::Value &value);
+  static pybind11::tuple args_to_tuple(const complate::Array &args);
+};
+
